refactor(CheckboxList): enhance state management with internal state handling
